+++ Frontend/src/containers/App.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import thunk from 'redux-thunk';
-import * as signalR from '@aspnet/signalr';
 import { createStore, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import { HashRouter, Route, Switch } from 'react-router-dom';
@@ -10,16 +9,6 @@ import { OpenRegistrationContainer } from '../registration/containers/OpenRegist
 import { reducer } from '../reducers';
 
 const store = createStore(reducer, applyMiddleware(thunk));
-
-let connection = new signalR.HubConnectionBuilder()
-    .withUrl("http://localhost:51474/ping")
-    .build();
- 
-connection.on("Notify", data => {
-    console.log(data);
-});
- 
-connection.start();
 
 export class App extends React.Component {
     render() {
